server/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "uvicorn>=0.40.0",
    "mypy>=1.19.1",
    "langchain-google-vertexai>=2.0.0",
    "langchain-core>=1.2.6",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "firebase-admin>=6.6.0",
    "google-auth>=2.0.0",
    "google-cloud-speech>=2.29.0",
    "google-cloud-secret-manager>=2.21.0",
    "langchain-groq>=1.1.1",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.14"
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = false
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "google.cloud.*"
ignore_missing_imports = true
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.infra.firestore_client"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.infra.repositories.firestore_session_repo"
ignore_errors = true

[[tool.mypy.overrides]]
module = "app.infra.repositories.firestore_conversation_repo"
ignore_errors = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.14.10",
]
